Binder_agent&image_search_agent/requirements.txt
crewai
crewai-tools

openai
azure-openai

azure-storage-blob
azure-identity
python-dotenv
Pillow


